feat(chat): parse sig messages
